Move api utils to testing/network package
